+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 2.1.56
+
+2022-05-26 10:20
+
+#### IMPROVED
+
+- In Markdown templates (erb), `i[:flagged]` will be set to true or false based on whether the item has @flagged (or whatever marker_tag is set to) present
+
+#### FIXED
+
+- `doing recent --interactive XX` wasn't respecting XX count and instead showing all
+- `doing recent --section SECT COUNT` not respecting count when section is provided
+
 ### 2.1.55
 
 2022-05-26 07:11